requirements.txt
discord.py
requests
Flask
python-dotenv
jokeapi